Allow for granular configuration of artifact generation and collection

+++ celery_workers/vv8_worker/tasks.py
@@ -20,7 +20,7 @@ def remove_entry(filepath):
 
 
 @celery_app.task(base=GridFSTask, bind=True, name='vv8_worker.process_url')
-def process_url(self, url: str, submission_id: str, mongo_id: str, crawler_args: List[str]):
+def process_url(self, url: str, submission_id: str, mongo_id: str, disable_har: bool, disable_screenshot: bool, disable_artifact_collection: bool, crawler_args: List[str]):
     print(f'vv8_worker process_url: url: {url}, submission_id: {submission_id}')
     start = time.perf_counter()
     crawler_path = os.path.join('/app', 'node/crawler.js')
@@ -51,21 +51,27 @@ def process_url(self, url: str, submission_id: str, mongo_id: str, crawler_args:
     self.update_state(state='PROGRESS', meta={'status': 'Uploading artifacts to mongodb'})
     screenshot_ids = []
     for screenshot in glob.glob("{}/*.png".format(wd_path)):
-        shutil.copy(screenshot, "/app/screenshots/{}".format(screenshot.split("/")[-1]))
-        file_id = self.gridfs.upload_from_stream(screenshot, open(screenshot, 'rb'), chunk_size_bytes=1024 * 1024, metadata={"contentType": "image/png"})
+        if not disable_screenshot:
+            shutil.copy(screenshot, "/app/screenshots/{}".format(screenshot.split("/")[-1]))
+            if ( not disable_artifact_collection ):
+                file_id = self.gridfs.upload_from_stream(screenshot, open(screenshot, 'rb'), chunk_size_bytes=1024 * 1024, metadata={"contentType": "image/png"})
+                screenshot_ids.append(file_id)
         os.remove(screenshot)
-        screenshot_ids.append(file_id)
     har_ids = []
     for har in glob.glob("{}/*.har".format(wd_path)):
-        shutil.copy(har, "/app/har/{}".format(har.split("/")[-1]))
-        file_id = self.gridfs.upload_from_stream(har, open(har, 'rb'), chunk_size_bytes=1024 * 1024, metadata={"contentType": "text/plain"})
-        os.remove(har)
-        har_ids.append(file_id)
+        if not disable_har:
+            shutil.copy(har, "/app/har/{}".format(har.split("/")[-1]))
+            if (not disable_artifact_collection):
+                file_id = self.gridfs.upload_from_stream(har, open(har, 'rb'), chunk_size_bytes=1024 * 1024, metadata={"contentType": "text/plain"})
+                har_ids.append(file_id)
+            os.remove(har)
     log_ids = []
     for entry in glob.glob(os.path.join(wd_path, 'vv8*.log')):
-        file_id = self.gridfs.upload_from_stream(entry, open(entry, 'rb'), chunk_size_bytes=1024 * 1024, metadata={"contentType": "text/plain"})
-        log_ids.append(file_id)
-    self.mongo['vv8_logs'].update_one({ '_id': ObjectId(mongo_id) }, { '$set': { 'screenshot_ids': screenshot_ids, 'har_ids': har_ids, 'log_ids': log_ids } })
+        if (not disable_artifact_collection):
+            file_id = self.gridfs.upload_from_stream(entry, open(entry, 'rb'), chunk_size_bytes=1024 * 1024, metadata={"contentType": "text/plain"})
+            log_ids.append(file_id)
+    if (not disable_artifact_collection):
+        self.mongo['vv8_logs'].update_one({ '_id': ObjectId(mongo_id) }, { '$set': { 'screenshot_ids': screenshot_ids, 'har_ids': har_ids, 'log_ids': log_ids } })
     if crawler_proc.returncode != 0:
         raise Exception('Crawler failed')
     end = time.perf_counter()
